I've changed a design of the app, add colors. Now daily menus are shown in different way - with use of Pager Adapter. Also the way of displaying meals in daily menus was changed. Now I'm using tags. They use new class ColorsBase to generate colors of tags that are compatible with background color of the app. Enabled adding, editing and deleting daily menus

+++ app/src/main/java/com/example/dell/menu/objects/Product.java
@@ -32,7 +32,7 @@ public class Product {
         return productId;
     }
 
-    public Product(String name, int quantity, String storageType){  //constructor for products that are to be showed in meals full activity
+    public Product(String name, double quantity, String storageType){  //constructor for products that are to be showed in meals full activity
         this.name = name;
         this.quantity = quantity;
         this.storageType = storageType;
